% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPerCellXeniumQCMetrics.R
\name{addPerCellXeniumQCMetrics}
\alias{addPerCellXeniumQCMetrics}
\title{A wrapper around \code{scuttle::addPerCellQCMetrics} to add Xenium-specific QC metrics based on negative control probe counts.}
\usage{
addPerCellXeniumQCMetrics(x)
}
\arguments{
\item{x}{a SpatialFeatureExperiment or SpatialExperiment object containing data from the Xenium In Situ platform}
}
\value{
x is returned with QC metrics added to \code{colData}
}
\description{
A wrapper around \code{scuttle::addPerCellQCMetrics} to add Xenium-specific QC metrics based on negative control probe counts.
}
