% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xenium.R
\name{readXenium}
\alias{readXenium}
\title{Read in Xenium data and create an SFE object}
\usage{
readXenium(dir_name)
}
\arguments{
\item{dir_name}{the name of the directory containing Xenium output files}
}
\value{
a SpatialFeatureExperiment object
}
\description{
Read in Xenium data and create an SFE object
}
