% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNegativeControlFeatures.R
\name{removeNegativeControlFeatures}
\alias{removeNegativeControlFeatures}
\title{Remove negative control probe and other technical features from Xenium In Situ data}
\usage{
removeNegativeControlFeatures(x)
}
\arguments{
\item{x}{a SpatialFeatureExperiment or SpatialExperiment object}
}
\value{
x, but only containing gene features
}
\description{
Remove negative control probe and other technical features from Xenium In Situ data
}
