% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFEtoSPE.R
\name{SFEtoSPE}
\alias{SFEtoSPE}
\title{Convert a SpatialFeatureExperiment object to a SpatialExperimentObject}
\usage{
SFEtoSPE(sfe)
}
\arguments{
\item{sfe}{a SpatialFeatureExperiment object}
}
\value{
a SpatialExperiment Object
}
\description{
Convert a SpatialFeatureExperiment object to a SpatialExperimentObject
}
